maintenance: Update Nest and TypeORM and introduce ESLint
NestJS updated from 8 to 10.

TypeORM from 0.2.45 to 0.3.20.

Warning: this is largely tested across the main features of the server when run with node. Also, the API has been largely tested from the client side API and all seems to be fine. But the testing is not complete, and it is not ready for deployment.

TypeORM removed a lot of syntactic shortcuts in their find and FindOne queries, which had been in use.

TypeORM configuration changes with respect to connections, which affected the configuration for the two databases.

TypeORM removed the ability to annotate a ManyToOne relationship as primary.  Consequently several "new" columns were defined to make foreign keys into primary keys as well.

TypORM deprecated Connection in favour of DataSource.

Lots of comments were added where it was perhaps unclear why the code is doing what it does. Mostly this was tennis related illumination.

Remove the PlayerIdentityService which was a lot of code meant to help TennisBC process their membership lists and identify which of their members had which VR Ids.  This went away in 2024 as Tennis BC elected to do their membership updates a different way.

Remove the ExternalApiModule which supported the PlayerIdentityService by using a public resource to look up common nicknames so things like Bob and Robert could be identified as potentially the same.

The AuthGuards I use all insert a User object into the express Request object and I used the resulting object without typing it.  So, I added a new Interface RequestWithUser to explicitly indicate wne the User object is available within the request.

SheetsJS (xlsx package) moved from npm repo to its own repo, which resolved some vulnerability warnings.

tournament grade approval was Not storing the approving user's name correctly.  Now it is.

Stopped using TSLint and started ESLint. Literally hundreds of tiny changes all done automatically then a few dozen that were good suggestions.